Initial GN checkin
work in progress, DON'T MERGE

main goal is to build everything with Visual Studio on Windows
android-emu-base and android-emu libraries build OK now on Linux
opengl stuff start to build with gn

+++ android/android-emu/android/base/memory/SharedMemory_posix.cpp
@@ -79,7 +79,7 @@ int SharedMemory::openInternal(int oflag, int mode) {
 
         // Only increase size, as we don't want to yank away memory
         // from another process.
-        if (mSize > sb.st_size && HANDLE_EINTR(ftruncate(mFd, mSize)) == -1) {
+        if (mSize > (size_t)sb.st_size && HANDLE_EINTR(ftruncate(mFd, mSize)) == -1) {
             err = -errno;
             close();
             return err;
